Add mvp of athlete detail page with results and upcoming entries

+++ app/src/main/java/ee/mtiidla/swimresult/data/network/mapper/AgeGroupNetworkMapper.kt
@@ -11,8 +11,8 @@ class AgeGroupNetworkMapper @Inject constructor() :
         AgeGroup(
             id = id.toLong(),
             key = key,
-            min = min,
-            max = max,
+            min = min.toInt(),
+            max = max.toInt(),
             name = name
         )
     }
